requirements.txt
freetype-py
pillow
numpy
